identifying posthog user once per page load
